+++ Changelog.md
@@ -2,6 +2,10 @@
 
 ## Gazebo 8.x.x (2017-xx-xx)
 
+1. Add the option --gui-client-plugin to load GUI plugins. Leave -g to load System Plugins.
+    * [Pull request]()
+    * [Issue 2279](https://bitbucket.org/osrf/gazebo/issues/2279)
+
 1. Add actors in World as models so they get returned with World::Models()
     * [Pull request #2706](https://bitbucket.org/osrf/gazebo/pull-request/2706)
     * [Issue #2271](https://bitbucket.org/osrf/gazebo/issues/2271)